MCR-3600  turn MCRXMLMetadataManager into an interface
