+++ mycore-base/src/main/resources/config/mycore.properties
@@ -288,7 +288,7 @@ MCR.Access.Facts.Condition.category=org.mycore.access.facts.condition.fact.MCRCa
   MCR.IFS2.SyncLastModifiedOnSVNCommit=true
 
 # Which metadata manager to use (dictates the available stores)
-  MCR.Metadata.Manager.Class=org.mycore.datamodel.common.MCRDefaultXMLMetadataManagerAdapter
+  MCR.Metadata.Manager.Class=org.mycore.datamodel.common.MCRDefaultXMLMetadataManager
 
 # Metadata store for derivate XML
   MCR.IFS2.Store.derivate.Class=org.mycore.datamodel.ifs2.MCRVersioningMetadataStore
