+++ mycore-base/src/test/java/org/mycore/datamodel/common/MCRXMLMetadataManagerTest.java
@@ -61,7 +61,7 @@ import org.xml.sax.SAXException;
 @MyCoReTest
 @ExtendWith(MCRMetadataExtension.class)
 @MCRTestConfiguration(properties = {
-    @MCRTestProperty(key = "MCR.Metadata.Manager.Class", classNameOf = MCRDefaultXMLMetadataManagerAdapter.class),
+    @MCRTestProperty(key = "MCR.Metadata.Manager.Class", classNameOf = MCRDefaultXMLMetadataManager.class),
     @MCRTestProperty(key = "MCR.Metadata.Type.document", string = "true"),
     @MCRTestProperty(key = "MCR.Metadata.ObjectID.NumberPattern", string = "00000000")
 })
