+++ mycore-ocfl/src/main/java/org/mycore/ocfl/commands/MCROCFLCommands.java
@@ -72,7 +72,7 @@ import org.mycore.frontend.cli.annotation.MCRCommand;
 import org.mycore.frontend.cli.annotation.MCRCommandGroup;
 import org.mycore.ocfl.classification.MCROCFLClassificationTransaction;
 import org.mycore.ocfl.classification.MCROCFLXMLClassificationManager;
-import org.mycore.ocfl.metadata.MCROCFLXMLMetadataManagerAdapter;
+import org.mycore.ocfl.metadata.MCROCFLXMLMetadataManager;
 import org.mycore.ocfl.metadata.migration.MCROCFLMigration;
 import org.mycore.ocfl.metadata.migration.MCROCFLRevisionPruner;
 import org.mycore.ocfl.niofs.MCROCFLFileSystemProvider;
@@ -133,8 +133,8 @@ public class MCROCFLCommands {
 
         MCROCFLMigration migration;
         if (metadataManagerConfigKey != null && !metadataManagerConfigKey.isEmpty()) {
-            MCROCFLXMLMetadataManagerAdapter metadataManager =
-                MCRConfiguration2.getInstanceOf(MCROCFLXMLMetadataManagerAdapter.class, metadataManagerConfigKey)
+            MCROCFLXMLMetadataManager metadataManager =
+                MCRConfiguration2.getInstanceOf(MCROCFLXMLMetadataManager.class, metadataManagerConfigKey)
                     .orElseThrow(() -> MCRConfiguration2.createConfigurationException(metadataManagerConfigKey));
             migration = new MCROCFLMigration(null, prunerList, metadataManager);
         } else if (repository != null && !repository.isEmpty()) {
@@ -326,7 +326,7 @@ public class MCROCFLCommands {
         help = "restore mcrobject {0} with version {1} to current store from ocfl history")
     public static void restoreObjFromOCFLVersioned(String mcridString, String revision) throws IOException {
         MCRObjectID mcrid = MCRObjectID.getInstance(mcridString);
-        MCROCFLXMLMetadataManagerAdapter manager = new MCROCFLXMLMetadataManagerAdapter();
+        MCROCFLXMLMetadataManager manager = new MCROCFLXMLMetadataManager();
         manager.setRepositoryKey(MCRConfiguration2.getStringOrThrow("MCR.Metadata.Manager.Repository"));
         MCRContent content = manager.retrieveContent(mcrid, revision);
         try {
@@ -361,7 +361,7 @@ public class MCROCFLCommands {
         help = "Permanently delete object {0} and its history from ocfl")
     public static void purgeObject(String mcridString) {
         MCRObjectID mcrid = MCRObjectID.getInstance(mcridString);
-        MCROCFLXMLMetadataManagerAdapter manager = new MCROCFLXMLMetadataManagerAdapter();
+        MCROCFLXMLMetadataManager manager = new MCROCFLXMLMetadataManager();
         manager.setRepositoryKey(MCRConfiguration2.getStringOrThrow("MCR.Metadata.Manager.Repository"));
         manager.purge(mcrid, new Date(), MCRUserManager.getCurrentUser().getUserName());
     }
@@ -410,7 +410,7 @@ public class MCROCFLCommands {
             return;
         }
         String repositoryKey = MCRConfiguration2.getStringOrThrow("MCR.Metadata.Manager.Repository");
-        MCROCFLXMLMetadataManagerAdapter manager = new MCROCFLXMLMetadataManagerAdapter();
+        MCROCFLXMLMetadataManager manager = new MCROCFLXMLMetadataManager();
         manager.setRepositoryKey(repositoryKey);
         OcflRepository repository = manager.getRepository();
         repository.listObjectIds()
