+++ mycore-ocfl/src/main/java/org/mycore/ocfl/metadata/MCRGZIPOCFLXMLMetadataManagerAdapter.java
@@ -18,46 +18,10 @@
 
 package org.mycore.ocfl.metadata;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.zip.GZIPInputStream;
-import java.util.zip.GZIPOutputStream;
+/**
+ * @deprecated Use {@link MCRGZIPOCFLXMLMetadataManager} instead.
+ */
+@Deprecated(forRemoval = true)
+public class MCRGZIPOCFLXMLMetadataManagerAdapter extends MCRGZIPOCFLXMLMetadataManager {
 
-import org.mycore.common.content.MCRContent;
-import org.mycore.datamodel.metadata.MCRObjectID;
-
-import io.ocfl.api.model.OcflObjectVersion;
-import io.ocfl.api.model.VersionNum;
-
-public class MCRGZIPOCFLXMLMetadataManagerAdapter extends MCROCFLXMLMetadataManagerAdapter {
-
-    @Override
-    protected InputStream getContentStream(MCRContent xml) throws IOException {
-        InputStream contentStream = super.getContentStream(xml);
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        try (GZIPOutputStream out = new GZIPOutputStream(baos)) {
-            contentStream.transferTo(out);
-        }
-        byte[] byteArray = baos.toByteArray();
-        return new ByteArrayInputStream(byteArray);
-    }
-
-    @Override
-    protected InputStream getStoredContentStream(MCRObjectID mcrid, OcflObjectVersion storeObject) throws IOException {
-        InputStream storedContentStream = super.getStoredContentStream(mcrid, storeObject);
-        return new GZIPInputStream(storedContentStream);
-    }
-
-    @Override
-    protected MCROCFLContent getContent(MCRObjectID id, String ocflObjectID, VersionNum key) {
-        return new MCRGZIPOCFLContent(getRepository(), ocflObjectID, buildFilePath(id),
-            key.toString());
-    }
-
-    @Override
-    protected String buildFilePath(MCRObjectID mcrid) {
-        return super.buildFilePath(mcrid) + ".gz";
-    }
 }
