+++ mycore-ocfl/src/test/resources/mycore.properties
@@ -7,7 +7,7 @@
 #####################
 
 MCR.Metadata.Manager.Repository=Test
-MCR.Metadata.Manager.Class=org.mycore.ocfl.metadata.MCROCFLXMLMetadataManagerAdapter
+MCR.Metadata.Manager.Class=org.mycore.ocfl.metadata.MCROCFLXMLMetadataManager
 
 #####################
 # NIO FS Properties
